Type Alias esp32c2::dma::out_conf0_ch::R

pub type R = R<OUT_CONF0_CH_SPEC>;

Register OUT_CONF0_CH%s reader

Aliased Type

struct R { /* private fields */ }

Implementations

impl R

pub fn out_rst(&self) -> OUT_RST_R

Bit 0 - This bit is used to reset DMA channel 0 Tx FSM and Tx FIFO pointer.

pub fn out_loop_test(&self) -> OUT_LOOP_TEST_R

Bit 1 - reserved

pub fn out_auto_wrback(&self) -> OUT_AUTO_WRBACK_R

Bit 2 - Set this bit to enable automatic outlink-writeback when all the data in tx buffer has been transmitted.

pub fn out_eof_mode(&self) -> OUT_EOF_MODE_R

Bit 3 - EOF flag generation mode when transmitting data. 1: EOF flag for Tx channel 0 is generated when data need to transmit has been popped from FIFO in DMA

pub fn outdscr_burst_en(&self) -> OUTDSCR_BURST_EN_R

Bit 4 - Set this bit to 1 to enable INCR burst transfer for Tx channel 0 reading link descriptor when accessing internal SRAM.

pub fn out_data_burst_en(&self) -> OUT_DATA_BURST_EN_R

Bit 5 - Set this bit to 1 to enable INCR burst transfer for Tx channel 0 transmitting data when accessing internal SRAM.