Module esp32c2::efuse::rd_repeat_data0

BLOCK0 data register 1.

Structs

RD_REPEAT_DATA0_SPEC

BLOCK0 data register 1.

Type Aliases

DIS_DIRECT_BOOT_R

Field DIS_DIRECT_BOOT reader - This bit set means disable direct_boot mode.

DIS_DOWNLOAD_ICACHE_R

Field DIS_DOWNLOAD_ICACHE reader - The bit be set to disable icache in download mode.

DIS_DOWNLOAD_MANUAL_ENCRYPT_R

Field DIS_DOWNLOAD_MANUAL_ENCRYPT reader - The bit be set to disable manual encryption.

DIS_DOWNLOAD_MODE_R

Field DIS_DOWNLOAD_MODE reader - Set this bit to disable download mode (boot_mode[3:0] = 0, 1, 2, 4, 5, 6, 7).

DIS_PAD_JTAG_R

Field DIS_PAD_JTAG reader - Set this bit to disable pad jtag.

ENABLE_SECURITY_DOWNLOAD_R

Field ENABLE_SECURITY_DOWNLOAD reader - Set this bit to enable secure UART download mode.

FLASH_TPUW_R

Field FLASH_TPUW reader - Configures flash waiting time after power-up, in unit of ms. If the value is less than 15, the waiting time is the configurable value. Otherwise, the waiting time is twice the configurable value.

FORCE_SEND_RESUME_R

Field FORCE_SEND_RESUME reader - Set this bit to force ROM code to send a resume command during SPI boot.

R

Register RD_REPEAT_DATA0 reader

RD_DIS_R

Field RD_DIS reader - The bit be set to disable software read high/low 128-bit of BLK3.

RPT4_RESERVED_R

Field RPT4_RESERVED reader - Reserved (used for four backups method).

SECURE_BOOT_EN_R

Field SECURE_BOOT_EN reader - The bit be set to enable secure boot.

SPI_BOOT_ENCRYPT_DECRYPT_CNT_R

Field SPI_BOOT_ENCRYPT_DECRYPT_CNT reader - These bits be set to enable SPI boot encrypt/decrypt. Odd number of 1: enable. even number of 1: disable.

UART_PRINT_CONTROL_R

Field UART_PRINT_CONTROL reader - Set this bit to disable usb printing.

WDT_DELAY_SEL_R

Field WDT_DELAY_SEL reader - Selects RTC watchdog timeout threshold, in unit of slow clock cycle. 0: 40000. 1: 80000. 2: 160000. 3:320000.

XTS_KEY_LENGTH_256_R

Field XTS_KEY_LENGTH_256 reader - The bit be set means XTS_AES use the whole 256-bit efuse data in BLOCK3. Otherwise, XTS_AES use 128-bit eFuse data in BLOCK3.