Type Alias esp32c2::efuse::rd_repeat_data0::R

pub type R = R<RD_REPEAT_DATA0_SPEC>;

Register RD_REPEAT_DATA0 reader

Aliased Type

struct R { /* private fields */ }

Implementations

impl R

pub fn rd_dis(&self) -> RD_DIS_R

Bits 0:1 - The bit be set to disable software read high/low 128-bit of BLK3.

pub fn wdt_delay_sel(&self) -> WDT_DELAY_SEL_R

Bits 2:3 - Selects RTC watchdog timeout threshold, in unit of slow clock cycle. 0: 40000. 1: 80000. 2: 160000. 3:320000.

pub fn dis_pad_jtag(&self) -> DIS_PAD_JTAG_R

Bit 4 - Set this bit to disable pad jtag.

pub fn dis_download_icache(&self) -> DIS_DOWNLOAD_ICACHE_R

Bit 5 - The bit be set to disable icache in download mode.

pub fn dis_download_manual_encrypt(&self) -> DIS_DOWNLOAD_MANUAL_ENCRYPT_R

Bit 6 - The bit be set to disable manual encryption.

pub fn spi_boot_encrypt_decrypt_cnt(&self) -> SPI_BOOT_ENCRYPT_DECRYPT_CNT_R

Bits 7:9 - These bits be set to enable SPI boot encrypt/decrypt. Odd number of 1: enable. even number of 1: disable.

pub fn xts_key_length_256(&self) -> XTS_KEY_LENGTH_256_R

Bit 10 - The bit be set means XTS_AES use the whole 256-bit efuse data in BLOCK3. Otherwise, XTS_AES use 128-bit eFuse data in BLOCK3.

pub fn uart_print_control(&self) -> UART_PRINT_CONTROL_R

Bits 11:12 - Set this bit to disable usb printing.

pub fn force_send_resume(&self) -> FORCE_SEND_RESUME_R

Bit 13 - Set this bit to force ROM code to send a resume command during SPI boot.

pub fn dis_download_mode(&self) -> DIS_DOWNLOAD_MODE_R

Bit 14 - Set this bit to disable download mode (boot_mode[3:0] = 0, 1, 2, 4, 5, 6, 7).

pub fn dis_direct_boot(&self) -> DIS_DIRECT_BOOT_R

Bit 15 - This bit set means disable direct_boot mode.

pub fn enable_security_download(&self) -> ENABLE_SECURITY_DOWNLOAD_R

Bit 16 - Set this bit to enable secure UART download mode.

pub fn flash_tpuw(&self) -> FLASH_TPUW_R

Bits 17:20 - Configures flash waiting time after power-up, in unit of ms. If the value is less than 15, the waiting time is the configurable value. Otherwise, the waiting time is twice the configurable value.

pub fn secure_boot_en(&self) -> SECURE_BOOT_EN_R

Bit 21 - The bit be set to enable secure boot.

pub fn rpt4_reserved(&self) -> RPT4_RESERVED_R

Bits 22:31 - Reserved (used for four backups method).