Module esp32c2::extmem

External Memory

Modules

cache_acs_cnt_clr

This description will be updated in the near future.

cache_conf_misc

This description will be updated in the near future.

cache_encrypt_decrypt_clk_force_on

This description will be updated in the near future.

cache_encrypt_decrypt_record_disable

This description will be updated in the near future.

cache_ilg_int_clr

This description will be updated in the near future.

cache_ilg_int_ena

This description will be updated in the near future.

cache_ilg_int_st

This description will be updated in the near future.

cache_mmu_fault_content

This description will be updated in the near future.

cache_mmu_fault_vaddr

This description will be updated in the near future.

cache_mmu_owner

This description will be updated in the near future.

cache_mmu_power_ctrl

This description will be updated in the near future.

cache_preload_int_ctrl

This description will be updated in the near future.

cache_request

This description will be updated in the near future.

cache_state

This description will be updated in the near future.

cache_sync_int_ctrl

This description will be updated in the near future.

cache_wrap_around_ctrl

This description will be updated in the near future.

clock_gate

This description will be updated in the near future.

core0_acs_cache_int_clr

This description will be updated in the near future.

core0_acs_cache_int_ena

This description will be updated in the near future.

core0_acs_cache_int_st

This description will be updated in the near future.

core0_dbus_reject_st

This description will be updated in the near future.

core0_dbus_reject_vaddr

This description will be updated in the near future.

core0_ibus_reject_st

This description will be updated in the near future.

core0_ibus_reject_vaddr

This description will be updated in the near future.

dbus_to_flash_end_vaddr

This description will be updated in the near future.

dbus_to_flash_start_vaddr

This description will be updated in the near future.

ibus_to_flash_end_vaddr

This description will be updated in the near future.

ibus_to_flash_start_vaddr

This description will be updated in the near future.

icache_atomic_operate_ena

This description will be updated in the near future.

icache_ctrl

This description will be updated in the near future.

icache_ctrl1

This description will be updated in the near future.

icache_freeze

This description will be updated in the near future.

icache_sync_addr

This description will be updated in the near future.

icache_sync_ctrl

This description will be updated in the near future.

icache_sync_size

This description will be updated in the near future.

icache_tag_power_ctrl

This description will be updated in the near future.

reg_date

This description will be updated in the near future.

Structs

RegisterBlock

Register block

Type Aliases

CACHE_ACS_CNT_CLR

CACHE_ACS_CNT_CLR (w) register accessor: This description will be updated in the near future.

CACHE_CONF_MISC

CACHE_CONF_MISC (rw) register accessor: This description will be updated in the near future.

CACHE_ENCRYPT_DECRYPT_CLK_FORCE_ON

CACHE_ENCRYPT_DECRYPT_CLK_FORCE_ON (rw) register accessor: This description will be updated in the near future.

CACHE_ENCRYPT_DECRYPT_RECORD_DISABLE

CACHE_ENCRYPT_DECRYPT_RECORD_DISABLE (rw) register accessor: This description will be updated in the near future.

CACHE_ILG_INT_CLR

CACHE_ILG_INT_CLR (w) register accessor: This description will be updated in the near future.

CACHE_ILG_INT_ENA

CACHE_ILG_INT_ENA (rw) register accessor: This description will be updated in the near future.

CACHE_ILG_INT_ST

CACHE_ILG_INT_ST (r) register accessor: This description will be updated in the near future.

CACHE_MMU_FAULT_CONTENT

CACHE_MMU_FAULT_CONTENT (r) register accessor: This description will be updated in the near future.

CACHE_MMU_FAULT_VADDR

CACHE_MMU_FAULT_VADDR (r) register accessor: This description will be updated in the near future.

CACHE_MMU_OWNER

CACHE_MMU_OWNER (rw) register accessor: This description will be updated in the near future.

CACHE_MMU_POWER_CTRL

CACHE_MMU_POWER_CTRL (rw) register accessor: This description will be updated in the near future.

CACHE_PRELOAD_INT_CTRL

CACHE_PRELOAD_INT_CTRL (rw) register accessor: This description will be updated in the near future.

CACHE_REQUEST

CACHE_REQUEST (rw) register accessor: This description will be updated in the near future.

CACHE_STATE

CACHE_STATE (r) register accessor: This description will be updated in the near future.

CACHE_SYNC_INT_CTRL

CACHE_SYNC_INT_CTRL (rw) register accessor: This description will be updated in the near future.

CACHE_WRAP_AROUND_CTRL

CACHE_WRAP_AROUND_CTRL (rw) register accessor: This description will be updated in the near future.

CLOCK_GATE

CLOCK_GATE (rw) register accessor: This description will be updated in the near future.

CORE0_ACS_CACHE_INT_CLR

CORE0_ACS_CACHE_INT_CLR (w) register accessor: This description will be updated in the near future.

CORE0_ACS_CACHE_INT_ENA

CORE0_ACS_CACHE_INT_ENA (rw) register accessor: This description will be updated in the near future.

CORE0_ACS_CACHE_INT_ST

CORE0_ACS_CACHE_INT_ST (r) register accessor: This description will be updated in the near future.

CORE0_DBUS_REJECT_ST

CORE0_DBUS_REJECT_ST (r) register accessor: This description will be updated in the near future.

CORE0_DBUS_REJECT_VADDR

CORE0_DBUS_REJECT_VADDR (r) register accessor: This description will be updated in the near future.

CORE0_IBUS_REJECT_ST

CORE0_IBUS_REJECT_ST (r) register accessor: This description will be updated in the near future.

CORE0_IBUS_REJECT_VADDR

CORE0_IBUS_REJECT_VADDR (r) register accessor: This description will be updated in the near future.

DBUS_TO_FLASH_END_VADDR

DBUS_TO_FLASH_END_VADDR (rw) register accessor: This description will be updated in the near future.

DBUS_TO_FLASH_START_VADDR

DBUS_TO_FLASH_START_VADDR (rw) register accessor: This description will be updated in the near future.

IBUS_TO_FLASH_END_VADDR

IBUS_TO_FLASH_END_VADDR (rw) register accessor: This description will be updated in the near future.

IBUS_TO_FLASH_START_VADDR

IBUS_TO_FLASH_START_VADDR (rw) register accessor: This description will be updated in the near future.

ICACHE_ATOMIC_OPERATE_ENA

ICACHE_ATOMIC_OPERATE_ENA (rw) register accessor: This description will be updated in the near future.

ICACHE_CTRL

ICACHE_CTRL (rw) register accessor: This description will be updated in the near future.

ICACHE_CTRL1

ICACHE_CTRL1 (rw) register accessor: This description will be updated in the near future.

ICACHE_FREEZE

ICACHE_FREEZE (rw) register accessor: This description will be updated in the near future.

ICACHE_SYNC_ADDR

ICACHE_SYNC_ADDR (rw) register accessor: This description will be updated in the near future.

ICACHE_SYNC_CTRL

ICACHE_SYNC_CTRL (rw) register accessor: This description will be updated in the near future.

ICACHE_SYNC_SIZE

ICACHE_SYNC_SIZE (rw) register accessor: This description will be updated in the near future.

ICACHE_TAG_POWER_CTRL

ICACHE_TAG_POWER_CTRL (rw) register accessor: This description will be updated in the near future.

REG_DATE

REG_DATE (rw) register accessor: This description will be updated in the near future.