Type Alias esp32c2::i2c0::ctr::R

pub type R = R<CTR_SPEC>;

Register CTR reader

Aliased Type

struct R { /* private fields */ }

Implementations

impl R

pub fn sda_force_out(&self) -> SDA_FORCE_OUT_R

Bit 0 - 0: direct output, 1: open drain output.

pub fn scl_force_out(&self) -> SCL_FORCE_OUT_R

Bit 1 - 0: direct output, 1: open drain output.

pub fn sample_scl_level(&self) -> SAMPLE_SCL_LEVEL_R

Bit 2 - This register is used to select the sample mode. 1: sample SDA data on the SCL low level. 0: sample SDA data on the SCL high level.

pub fn rx_full_ack_level(&self) -> RX_FULL_ACK_LEVEL_R

Bit 3 - This register is used to configure the ACK value that need to sent by master when the rx_fifo_cnt has reached the threshold.

pub fn ms_mode(&self) -> MS_MODE_R

Bit 4 - Set this bit to configure the module as an I2C Master. Clear this bit to configure the module as an I2C Slave.

pub fn tx_lsb_first(&self) -> TX_LSB_FIRST_R

Bit 6 - This bit is used to control the sending mode for data needing to be sent. 1: send data from the least significant bit, 0: send data from the most significant bit.

pub fn rx_lsb_first(&self) -> RX_LSB_FIRST_R

Bit 7 - This bit is used to control the storage mode for received data. 1: receive data from the least significant bit, 0: receive data from the most significant bit.

pub fn clk_en(&self) -> CLK_EN_R

Bit 8 - Reserved

pub fn arbitration_en(&self) -> ARBITRATION_EN_R

Bit 9 - This is the enable bit for arbitration_lost.

pub fn slv_tx_auto_start_en(&self) -> SLV_TX_AUTO_START_EN_R

Bit 12 - This is the enable bit for slave to send data automatically