Type Alias esp32c2::ledc::timer_conf::R

pub type R = R<TIMER_CONF_SPEC>;

Register TIMER%s_CONF reader

Aliased Type

struct R { /* private fields */ }

Implementations

impl R

pub fn duty_res(&self) -> DUTY_RES_R

Bits 0:3 - This register is used to control the range of the counter in timer %s.

pub fn clk_div(&self) -> CLK_DIV_R

Bits 4:21 - This register is used to configure the divisor for the divider in timer %s. The least significant eight bits represent the fractional part.

pub fn pause(&self) -> PAUSE_R

Bit 22 - This bit is used to suspend the counter in timer %s.

pub fn rst(&self) -> RST_R

Bit 23 - This bit is used to reset timer %s. The counter will show 0 after reset.

pub fn tick_sel(&self) -> TICK_SEL_R

Bit 24 - This bit is used to select clock for timer %s. When this bit is set to 1 LEDC_APB_CLK_SEL[1:0] should be 1, otherwise the timer clock may be not accurate. 1’h0: SLOW_CLK 1’h1: REF_TICK