Module esp32c2::rtc_cntl::low_power_st

register description

Structs

LOW_POWER_ST_SPEC

register description

Type Aliases

COCPU_STATE_DONE_R

Field COCPU_STATE_DONE reader - ulp/cocpu is done

COCPU_STATE_DONE_W

Field COCPU_STATE_DONE writer - ulp/cocpu is done

COCPU_STATE_SLP_R

Field COCPU_STATE_SLP reader - ulp/cocpu is in sleep state

COCPU_STATE_SLP_W

Field COCPU_STATE_SLP writer - ulp/cocpu is in sleep state

COCPU_STATE_START_R

Field COCPU_STATE_START reader - ulp/cocpu should start to work

COCPU_STATE_START_W

Field COCPU_STATE_START writer - ulp/cocpu should start to work

COCPU_STATE_SWITCH_R

Field COCPU_STATE_SWITCH reader - ulp/cocpu is about to working. Switch rtc main state

COCPU_STATE_SWITCH_W

Field COCPU_STATE_SWITCH writer - ulp/cocpu is about to working. Switch rtc main state

IN_LOW_POWER_STATE_R

Field IN_LOW_POWER_STATE reader - rtc main state machine is in the states of low power

IN_LOW_POWER_STATE_W

Field IN_LOW_POWER_STATE writer - rtc main state machine is in the states of low power

IN_WAKEUP_STATE_R

Field IN_WAKEUP_STATE reader - rtc main state machine is in the states of wakeup process

IN_WAKEUP_STATE_W

Field IN_WAKEUP_STATE writer - rtc main state machine is in the states of wakeup process

MAIN_STATE_IN_IDLE_R

Field MAIN_STATE_IN_IDLE reader - rtc main state machine is in idle state

MAIN_STATE_IN_IDLE_W

Field MAIN_STATE_IN_IDLE writer - rtc main state machine is in idle state

MAIN_STATE_IN_SLP_R

Field MAIN_STATE_IN_SLP reader - rtc main state machine is in sleep state

MAIN_STATE_IN_SLP_W

Field MAIN_STATE_IN_SLP writer - rtc main state machine is in sleep state

MAIN_STATE_IN_WAIT_8M_R

Field MAIN_STATE_IN_WAIT_8M reader - rtc main state machine is in wait 8m state

MAIN_STATE_IN_WAIT_8M_W

Field MAIN_STATE_IN_WAIT_8M writer - rtc main state machine is in wait 8m state

MAIN_STATE_IN_WAIT_PLL_R

Field MAIN_STATE_IN_WAIT_PLL reader - rtc main state machine is in wait pll state

MAIN_STATE_IN_WAIT_PLL_W

Field MAIN_STATE_IN_WAIT_PLL writer - rtc main state machine is in wait pll state

MAIN_STATE_IN_WAIT_XTL_R

Field MAIN_STATE_IN_WAIT_XTL reader - rtc main state machine is in wait xtal state

MAIN_STATE_IN_WAIT_XTL_W

Field MAIN_STATE_IN_WAIT_XTL writer - rtc main state machine is in wait xtal state

MAIN_STATE_PLL_ON_R

Field MAIN_STATE_PLL_ON reader - rtc main state machine is in states that pll should be running

MAIN_STATE_PLL_ON_W

Field MAIN_STATE_PLL_ON writer - rtc main state machine is in states that pll should be running

MAIN_STATE_R

Field MAIN_STATE reader - rtc main state machine status

MAIN_STATE_W

Field MAIN_STATE writer - rtc main state machine status

MAIN_STATE_WAIT_END_R

Field MAIN_STATE_WAIT_END reader - rtc main state machine has been waited for some cycles

MAIN_STATE_WAIT_END_W

Field MAIN_STATE_WAIT_END writer - rtc main state machine has been waited for some cycles

MAIN_STATE_XTAL_ISO_R

Field MAIN_STATE_XTAL_ISO reader - no use any more

MAIN_STATE_XTAL_ISO_W

Field MAIN_STATE_XTAL_ISO writer - no use any more

R

Register LOW_POWER_ST reader

RDY_FOR_WAKEUP_R

Field RDY_FOR_WAKEUP reader - rtc is ready to receive wake up trigger from wake up source

RDY_FOR_WAKEUP_W

Field RDY_FOR_WAKEUP writer - rtc is ready to receive wake up trigger from wake up source

TOUCH_STATE_DONE_R

Field TOUCH_STATE_DONE reader - touch is done

TOUCH_STATE_DONE_W

Field TOUCH_STATE_DONE writer - touch is done

TOUCH_STATE_SLP_R

Field TOUCH_STATE_SLP reader - touch is in sleep state

TOUCH_STATE_SLP_W

Field TOUCH_STATE_SLP writer - touch is in sleep state

TOUCH_STATE_START_R

Field TOUCH_STATE_START reader - touch should start to work

TOUCH_STATE_START_W

Field TOUCH_STATE_START writer - touch should start to work

TOUCH_STATE_SWITCH_R

Field TOUCH_STATE_SWITCH reader - touch is about to working. Switch rtc main state

TOUCH_STATE_SWITCH_W

Field TOUCH_STATE_SWITCH writer - touch is about to working. Switch rtc main state

W

Register LOW_POWER_ST writer

XPD_DIG_R

Field XPD_DIG reader - digital wrap power down

XPD_DIG_W

Field XPD_DIG writer - digital wrap power down