Type Alias esp32c2::rtc_cntl::low_power_st::R

pub type R = R<LOW_POWER_ST_SPEC>;

Register LOW_POWER_ST reader

Aliased Type

struct R { /* private fields */ }

Implementations

impl R

pub fn xpd_dig(&self) -> XPD_DIG_R

Bit 8 - digital wrap power down

pub fn touch_state_start(&self) -> TOUCH_STATE_START_R

Bit 9 - touch should start to work

pub fn touch_state_switch(&self) -> TOUCH_STATE_SWITCH_R

Bit 10 - touch is about to working. Switch rtc main state

pub fn touch_state_slp(&self) -> TOUCH_STATE_SLP_R

Bit 11 - touch is in sleep state

pub fn touch_state_done(&self) -> TOUCH_STATE_DONE_R

Bit 12 - touch is done

pub fn cocpu_state_start(&self) -> COCPU_STATE_START_R

Bit 13 - ulp/cocpu should start to work

pub fn cocpu_state_switch(&self) -> COCPU_STATE_SWITCH_R

Bit 14 - ulp/cocpu is about to working. Switch rtc main state

pub fn cocpu_state_slp(&self) -> COCPU_STATE_SLP_R

Bit 15 - ulp/cocpu is in sleep state

pub fn cocpu_state_done(&self) -> COCPU_STATE_DONE_R

Bit 16 - ulp/cocpu is done

pub fn main_state_xtal_iso(&self) -> MAIN_STATE_XTAL_ISO_R

Bit 17 - no use any more

pub fn main_state_pll_on(&self) -> MAIN_STATE_PLL_ON_R

Bit 18 - rtc main state machine is in states that pll should be running

pub fn rdy_for_wakeup(&self) -> RDY_FOR_WAKEUP_R

Bit 19 - rtc is ready to receive wake up trigger from wake up source

pub fn main_state_wait_end(&self) -> MAIN_STATE_WAIT_END_R

Bit 20 - rtc main state machine has been waited for some cycles

pub fn in_wakeup_state(&self) -> IN_WAKEUP_STATE_R

Bit 21 - rtc main state machine is in the states of wakeup process

pub fn in_low_power_state(&self) -> IN_LOW_POWER_STATE_R

Bit 22 - rtc main state machine is in the states of low power

pub fn main_state_in_wait_8m(&self) -> MAIN_STATE_IN_WAIT_8M_R

Bit 23 - rtc main state machine is in wait 8m state

pub fn main_state_in_wait_pll(&self) -> MAIN_STATE_IN_WAIT_PLL_R

Bit 24 - rtc main state machine is in wait pll state

pub fn main_state_in_wait_xtl(&self) -> MAIN_STATE_IN_WAIT_XTL_R

Bit 25 - rtc main state machine is in wait xtal state

pub fn main_state_in_slp(&self) -> MAIN_STATE_IN_SLP_R

Bit 26 - rtc main state machine is in sleep state

pub fn main_state_in_idle(&self) -> MAIN_STATE_IN_IDLE_R

Bit 27 - rtc main state machine is in idle state

pub fn main_state(&self) -> MAIN_STATE_R

Bits 28:31 - rtc main state machine status