Type Alias esp32c2::rtc_cntl::low_power_st::W

pub type W = W<LOW_POWER_ST_SPEC>;

Register LOW_POWER_ST writer

Aliased Type

struct W { /* private fields */ }

Implementations

impl W

pub fn xpd_dig(&mut self) -> XPD_DIG_W<'_, LOW_POWER_ST_SPEC>

Bit 8 - digital wrap power down

pub fn touch_state_start( &mut self ) -> TOUCH_STATE_START_W<'_, LOW_POWER_ST_SPEC>

Bit 9 - touch should start to work

pub fn touch_state_switch( &mut self ) -> TOUCH_STATE_SWITCH_W<'_, LOW_POWER_ST_SPEC>

Bit 10 - touch is about to working. Switch rtc main state

pub fn touch_state_slp(&mut self) -> TOUCH_STATE_SLP_W<'_, LOW_POWER_ST_SPEC>

Bit 11 - touch is in sleep state

pub fn touch_state_done(&mut self) -> TOUCH_STATE_DONE_W<'_, LOW_POWER_ST_SPEC>

Bit 12 - touch is done

pub fn cocpu_state_start( &mut self ) -> COCPU_STATE_START_W<'_, LOW_POWER_ST_SPEC>

Bit 13 - ulp/cocpu should start to work

pub fn cocpu_state_switch( &mut self ) -> COCPU_STATE_SWITCH_W<'_, LOW_POWER_ST_SPEC>

Bit 14 - ulp/cocpu is about to working. Switch rtc main state

pub fn cocpu_state_slp(&mut self) -> COCPU_STATE_SLP_W<'_, LOW_POWER_ST_SPEC>

Bit 15 - ulp/cocpu is in sleep state

pub fn cocpu_state_done(&mut self) -> COCPU_STATE_DONE_W<'_, LOW_POWER_ST_SPEC>

Bit 16 - ulp/cocpu is done

pub fn main_state_xtal_iso( &mut self ) -> MAIN_STATE_XTAL_ISO_W<'_, LOW_POWER_ST_SPEC>

Bit 17 - no use any more

pub fn main_state_pll_on( &mut self ) -> MAIN_STATE_PLL_ON_W<'_, LOW_POWER_ST_SPEC>

Bit 18 - rtc main state machine is in states that pll should be running

pub fn rdy_for_wakeup(&mut self) -> RDY_FOR_WAKEUP_W<'_, LOW_POWER_ST_SPEC>

Bit 19 - rtc is ready to receive wake up trigger from wake up source

pub fn main_state_wait_end( &mut self ) -> MAIN_STATE_WAIT_END_W<'_, LOW_POWER_ST_SPEC>

Bit 20 - rtc main state machine has been waited for some cycles

pub fn in_wakeup_state(&mut self) -> IN_WAKEUP_STATE_W<'_, LOW_POWER_ST_SPEC>

Bit 21 - rtc main state machine is in the states of wakeup process

pub fn in_low_power_state( &mut self ) -> IN_LOW_POWER_STATE_W<'_, LOW_POWER_ST_SPEC>

Bit 22 - rtc main state machine is in the states of low power

pub fn main_state_in_wait_8m( &mut self ) -> MAIN_STATE_IN_WAIT_8M_W<'_, LOW_POWER_ST_SPEC>

Bit 23 - rtc main state machine is in wait 8m state

pub fn main_state_in_wait_pll( &mut self ) -> MAIN_STATE_IN_WAIT_PLL_W<'_, LOW_POWER_ST_SPEC>

Bit 24 - rtc main state machine is in wait pll state

pub fn main_state_in_wait_xtl( &mut self ) -> MAIN_STATE_IN_WAIT_XTL_W<'_, LOW_POWER_ST_SPEC>

Bit 25 - rtc main state machine is in wait xtal state

pub fn main_state_in_slp( &mut self ) -> MAIN_STATE_IN_SLP_W<'_, LOW_POWER_ST_SPEC>

Bit 26 - rtc main state machine is in sleep state

pub fn main_state_in_idle( &mut self ) -> MAIN_STATE_IN_IDLE_W<'_, LOW_POWER_ST_SPEC>

Bit 27 - rtc main state machine is in idle state

pub fn main_state(&mut self) -> MAIN_STATE_W<'_, LOW_POWER_ST_SPEC>

Bits 28:31 - rtc main state machine status

pub unsafe fn bits(&mut self, bits: u32) -> &mut Self

Writes raw bits to the register.

Safety

Passing incorrect value can cause undefined behaviour. See reference manual