Module esp32c2::spi0::ctrl

SPI0 control register.

Structs

CTRL_SPEC

SPI0 control register.

Type Aliases

D_POL_R

Field D_POL reader - The bit is used to set MOSI line polarity, 1: high 0, low

D_POL_W

Field D_POL writer - The bit is used to set MOSI line polarity, 1: high 0, low

FASTRD_MODE_R

Field FASTRD_MODE reader - This bit enable the bits: SPI_MEM_FREAD_QIO, SPI_MEM_FREAD_DIO, SPI_MEM_FREAD_QOUT AND SPI_MEM_FREAD_DOUT. 1: enable 0: disable.

FASTRD_MODE_W

Field FASTRD_MODE writer - This bit enable the bits: SPI_MEM_FREAD_QIO, SPI_MEM_FREAD_DIO, SPI_MEM_FREAD_QOUT AND SPI_MEM_FREAD_DOUT. 1: enable 0: disable.

FCMD_DUAL_R

Field FCMD_DUAL reader - Apply 2 signals during command phase 1:enable 0: disable

FCMD_DUAL_W

Field FCMD_DUAL writer - Apply 2 signals during command phase 1:enable 0: disable

FCMD_QUAD_R

Field FCMD_QUAD reader - Apply 4 signals during command phase 1:enable 0: disable

FCMD_QUAD_W

Field FCMD_QUAD writer - Apply 4 signals during command phase 1:enable 0: disable

FDUMMY_OUT_R

Field FDUMMY_OUT reader - In the dummy phase the signal level of spi is output by the spi controller.

FDUMMY_OUT_W

Field FDUMMY_OUT writer - In the dummy phase the signal level of spi is output by the spi controller.

FREAD_DIO_R

Field FREAD_DIO reader - In the read operations address phase and read-data phase apply 2 signals. 1: enable 0: disable.

FREAD_DIO_W

Field FREAD_DIO writer - In the read operations address phase and read-data phase apply 2 signals. 1: enable 0: disable.

FREAD_DUAL_R

Field FREAD_DUAL reader - In the read operations, read-data phase apply 2 signals. 1: enable 0: disable.

FREAD_DUAL_W

Field FREAD_DUAL writer - In the read operations, read-data phase apply 2 signals. 1: enable 0: disable.

FREAD_QIO_R

Field FREAD_QIO reader - In the read operations address phase and read-data phase apply 4 signals. 1: enable 0: disable.

FREAD_QIO_W

Field FREAD_QIO writer - In the read operations address phase and read-data phase apply 4 signals. 1: enable 0: disable.

FREAD_QUAD_R

Field FREAD_QUAD reader - In the read operations read-data phase apply 4 signals. 1: enable 0: disable.

FREAD_QUAD_W

Field FREAD_QUAD writer - In the read operations read-data phase apply 4 signals. 1: enable 0: disable.

Q_POL_R

Field Q_POL reader - The bit is used to set MISO line polarity, 1: high 0, low

Q_POL_W

Field Q_POL writer - The bit is used to set MISO line polarity, 1: high 0, low

R

Register CTRL reader

W

Register CTRL writer

WP_R

Field WP reader - Write protect signal output when SPI is idle. 1: output high, 0: output low.

WP_W

Field WP writer - Write protect signal output when SPI is idle. 1: output high, 0: output low.