Module esp32c2::spi0::misc

SPI0 misc register

Structs

MISC_SPEC

SPI0 misc register

Type Aliases

CK_IDLE_EDGE_R

Field CK_IDLE_EDGE reader - 1: spi clk line is high when idle 0: spi clk line is low when idle

CK_IDLE_EDGE_W

Field CK_IDLE_EDGE writer - 1: spi clk line is high when idle 0: spi clk line is low when idle

CSPI_ST_TRANS_END_INT_ENA_R

Field CSPI_ST_TRANS_END_INT_ENA reader - The bit is used to enable the interrupt of spi0_slv_st controlled transmitting is done.

CSPI_ST_TRANS_END_INT_ENA_W

Field CSPI_ST_TRANS_END_INT_ENA writer - The bit is used to enable the interrupt of spi0_slv_st controlled transmitting is done.

CSPI_ST_TRANS_END_R

Field CSPI_ST_TRANS_END reader - The bit is used to indicate the spi0_slv_st controlled transmitting is done.

CSPI_ST_TRANS_END_W

Field CSPI_ST_TRANS_END writer - The bit is used to indicate the spi0_slv_st controlled transmitting is done.

CS_KEEP_ACTIVE_R

Field CS_KEEP_ACTIVE reader - spi cs line keep low when the bit is set.

CS_KEEP_ACTIVE_W

Field CS_KEEP_ACTIVE writer - spi cs line keep low when the bit is set.

R

Register MISC reader

TRANS_END_INT_ENA_R

Field TRANS_END_INT_ENA reader - The bit is used to enable the interrupt of spi0_mst_st controlled transmitting is done.

TRANS_END_INT_ENA_W

Field TRANS_END_INT_ENA writer - The bit is used to enable the interrupt of spi0_mst_st controlled transmitting is done.

TRANS_END_R

Field TRANS_END reader - The bit is used to indicate the spi0_mst_st controlled transmitting is done.

TRANS_END_W

Field TRANS_END writer - The bit is used to indicate the spi0_mst_st controlled transmitting is done.

W

Register MISC writer