Module esp32c2::spi1::user

SPI1 user register.

Structs

USER_SPEC

SPI1 user register.

Type Aliases

CK_OUT_EDGE_R

Field CK_OUT_EDGE reader - the bit combined with spi_mem_mosi_delay_mode bits to set mosi signal delay mode.

CK_OUT_EDGE_W

Field CK_OUT_EDGE writer - the bit combined with spi_mem_mosi_delay_mode bits to set mosi signal delay mode.

FWRITE_DIO_R

Field FWRITE_DIO reader - In the write operations address phase and read-data phase apply 2 signals.

FWRITE_DIO_W

Field FWRITE_DIO writer - In the write operations address phase and read-data phase apply 2 signals.

FWRITE_DUAL_R

Field FWRITE_DUAL reader - In the write operations read-data phase apply 2 signals

FWRITE_DUAL_W

Field FWRITE_DUAL writer - In the write operations read-data phase apply 2 signals

FWRITE_QIO_R

Field FWRITE_QIO reader - In the write operations address phase and read-data phase apply 4 signals.

FWRITE_QIO_W

Field FWRITE_QIO writer - In the write operations address phase and read-data phase apply 4 signals.

FWRITE_QUAD_R

Field FWRITE_QUAD reader - In the write operations read-data phase apply 4 signals

FWRITE_QUAD_W

Field FWRITE_QUAD writer - In the write operations read-data phase apply 4 signals

R

Register USER reader

USR_ADDR_R

Field USR_ADDR reader - This bit enable the address phase of an operation.

USR_ADDR_W

Field USR_ADDR writer - This bit enable the address phase of an operation.

USR_COMMAND_R

Field USR_COMMAND reader - This bit enable the command phase of an operation.

USR_COMMAND_W

Field USR_COMMAND writer - This bit enable the command phase of an operation.

USR_DUMMY_IDLE_R

Field USR_DUMMY_IDLE reader - SPI clock is disable in dummy phase when the bit is enable.

USR_DUMMY_IDLE_W

Field USR_DUMMY_IDLE writer - SPI clock is disable in dummy phase when the bit is enable.

USR_DUMMY_R

Field USR_DUMMY reader - This bit enable the dummy phase of an operation.

USR_DUMMY_W

Field USR_DUMMY writer - This bit enable the dummy phase of an operation.

USR_MISO_HIGHPART_R

Field USR_MISO_HIGHPART reader - read-data phase only access to high-part of the buffer spi_mem_w8~spi_mem_w15. 1: enable 0: disable.

USR_MISO_HIGHPART_W

Field USR_MISO_HIGHPART writer - read-data phase only access to high-part of the buffer spi_mem_w8~spi_mem_w15. 1: enable 0: disable.

USR_MISO_R

Field USR_MISO reader - This bit enable the read-data phase of an operation.

USR_MISO_W

Field USR_MISO writer - This bit enable the read-data phase of an operation.

USR_MOSI_HIGHPART_R

Field USR_MOSI_HIGHPART reader - write-data phase only access to high-part of the buffer spi_mem_w8~spi_mem_w15. 1: enable 0: disable.

USR_MOSI_HIGHPART_W

Field USR_MOSI_HIGHPART writer - write-data phase only access to high-part of the buffer spi_mem_w8~spi_mem_w15. 1: enable 0: disable.

USR_MOSI_R

Field USR_MOSI reader - This bit enable the write-data phase of an operation.

USR_MOSI_W

Field USR_MOSI writer - This bit enable the write-data phase of an operation.

W

Register USER writer