Type Alias esp32c2::spi2::clock::R

pub type R = R<CLOCK_SPEC>;

Register CLOCK reader

Aliased Type

struct R { /* private fields */ }

Implementations

impl R

pub fn clkcnt_l(&self) -> CLKCNT_L_R

Bits 0:5 - In the master mode it must be equal to spi_clkcnt_N. In the slave mode it must be 0. Can be configured in CONF state.

pub fn clkcnt_h(&self) -> CLKCNT_H_R

Bits 6:11 - In the master mode it must be floor((spi_clkcnt_N+1)/2-1). In the slave mode it must be 0. Can be configured in CONF state.

pub fn clkcnt_n(&self) -> CLKCNT_N_R

Bits 12:17 - In the master mode it is the divider of spi_clk. So spi_clk frequency is system/(spi_clkdiv_pre+1)/(spi_clkcnt_N+1). Can be configured in CONF state.

pub fn clkdiv_pre(&self) -> CLKDIV_PRE_R

Bits 18:21 - In the master mode it is pre-divider of spi_clk. Can be configured in CONF state.

pub fn clk_equ_sysclk(&self) -> CLK_EQU_SYSCLK_R

Bit 31 - In the master mode 1: spi_clk is eqaul to system 0: spi_clk is divided from system clock. Can be configured in CONF state.