Module esp32c2::spi2::dma_int_raw

SPI interrupt raw register

Structs

DMA_INT_RAW_SPEC

SPI interrupt raw register

Type Aliases

APP1_INT_RAW_R

Field APP1_INT_RAW reader - The raw bit for SPI_APP1_INT interrupt. The value is only controlled by software.

APP1_INT_RAW_W

Field APP1_INT_RAW writer - The raw bit for SPI_APP1_INT interrupt. The value is only controlled by software.

APP2_INT_RAW_R

Field APP2_INT_RAW reader - The raw bit for SPI_APP2_INT interrupt. The value is only controlled by software.

APP2_INT_RAW_W

Field APP2_INT_RAW writer - The raw bit for SPI_APP2_INT interrupt. The value is only controlled by software.

DMA_INFIFO_FULL_ERR_INT_RAW_R

Field DMA_INFIFO_FULL_ERR_INT_RAW reader - 1: The current data rate of DMA Rx is smaller than that of SPI, which will lose the receive data. 0: Others.

DMA_INFIFO_FULL_ERR_INT_RAW_W

Field DMA_INFIFO_FULL_ERR_INT_RAW writer - 1: The current data rate of DMA Rx is smaller than that of SPI, which will lose the receive data. 0: Others.

DMA_OUTFIFO_EMPTY_ERR_INT_RAW_R

Field DMA_OUTFIFO_EMPTY_ERR_INT_RAW reader - 1: The current data rate of DMA TX is smaller than that of SPI. SPI will stop in master mode and send out all 0 in slave mode. 0: Others.

DMA_OUTFIFO_EMPTY_ERR_INT_RAW_W

Field DMA_OUTFIFO_EMPTY_ERR_INT_RAW writer - 1: The current data rate of DMA TX is smaller than that of SPI. SPI will stop in master mode and send out all 0 in slave mode. 0: Others.

DMA_SEG_TRANS_DONE_INT_RAW_R

Field DMA_SEG_TRANS_DONE_INT_RAW reader - The raw bit for SPI_DMA_SEG_TRANS_DONE_INT interrupt. 1: spi master DMA full-duplex/half-duplex seg-conf-trans ends or slave half-duplex seg-trans ends. And data has been pushed to corresponding memory. 0: seg-conf-trans or seg-trans is not ended or not occurred.

DMA_SEG_TRANS_DONE_INT_RAW_W

Field DMA_SEG_TRANS_DONE_INT_RAW writer - The raw bit for SPI_DMA_SEG_TRANS_DONE_INT interrupt. 1: spi master DMA full-duplex/half-duplex seg-conf-trans ends or slave half-duplex seg-trans ends. And data has been pushed to corresponding memory. 0: seg-conf-trans or seg-trans is not ended or not occurred.

MST_RX_AFIFO_WFULL_ERR_INT_RAW_R

Field MST_RX_AFIFO_WFULL_ERR_INT_RAW reader - The raw bit for SPI_MST_RX_AFIFO_WFULL_ERR_INT interrupt. 1: There is a RX AFIFO write-full error when SPI inputs data in master mode. 0: Others.

MST_RX_AFIFO_WFULL_ERR_INT_RAW_W

Field MST_RX_AFIFO_WFULL_ERR_INT_RAW writer - The raw bit for SPI_MST_RX_AFIFO_WFULL_ERR_INT interrupt. 1: There is a RX AFIFO write-full error when SPI inputs data in master mode. 0: Others.

MST_TX_AFIFO_REMPTY_ERR_INT_RAW_R

Field MST_TX_AFIFO_REMPTY_ERR_INT_RAW reader - The raw bit for SPI_MST_TX_AFIFO_REMPTY_ERR_INT interrupt. 1: There is a TX BUF AFIFO read-empty error when SPI outputs data in master mode. 0: Others.

MST_TX_AFIFO_REMPTY_ERR_INT_RAW_W

Field MST_TX_AFIFO_REMPTY_ERR_INT_RAW writer - The raw bit for SPI_MST_TX_AFIFO_REMPTY_ERR_INT interrupt. 1: There is a TX BUF AFIFO read-empty error when SPI outputs data in master mode. 0: Others.

R

Register DMA_INT_RAW reader

SEG_MAGIC_ERR_INT_RAW_R

Field SEG_MAGIC_ERR_INT_RAW reader - The raw bit for SPI_SEG_MAGIC_ERR_INT interrupt. 1: The magic value in CONF buffer is error in the DMA seg-conf-trans. 0: others.

SEG_MAGIC_ERR_INT_RAW_W

Field SEG_MAGIC_ERR_INT_RAW writer - The raw bit for SPI_SEG_MAGIC_ERR_INT interrupt. 1: The magic value in CONF buffer is error in the DMA seg-conf-trans. 0: others.

SLV_BUF_ADDR_ERR_INT_RAW_R

Field SLV_BUF_ADDR_ERR_INT_RAW reader - The raw bit for SPI_SLV_BUF_ADDR_ERR_INT interrupt. 1: The accessing data address of the current SPI slave mode CPU controlled FD, Wr_BUF or Rd_BUF transmission is bigger than 63. 0: Others.

SLV_BUF_ADDR_ERR_INT_RAW_W

Field SLV_BUF_ADDR_ERR_INT_RAW writer - The raw bit for SPI_SLV_BUF_ADDR_ERR_INT interrupt. 1: The accessing data address of the current SPI slave mode CPU controlled FD, Wr_BUF or Rd_BUF transmission is bigger than 63. 0: Others.

SLV_CMD7_INT_RAW_R

Field SLV_CMD7_INT_RAW reader - The raw bit for SPI slave CMD7 interrupt. 1: SPI slave mode CMD7 transmission is ended. 0: Others.

SLV_CMD7_INT_RAW_W

Field SLV_CMD7_INT_RAW writer - The raw bit for SPI slave CMD7 interrupt. 1: SPI slave mode CMD7 transmission is ended. 0: Others.

SLV_CMD8_INT_RAW_R

Field SLV_CMD8_INT_RAW reader - The raw bit for SPI slave CMD8 interrupt. 1: SPI slave mode CMD8 transmission is ended. 0: Others.

SLV_CMD8_INT_RAW_W

Field SLV_CMD8_INT_RAW writer - The raw bit for SPI slave CMD8 interrupt. 1: SPI slave mode CMD8 transmission is ended. 0: Others.

SLV_CMD9_INT_RAW_R

Field SLV_CMD9_INT_RAW reader - The raw bit for SPI slave CMD9 interrupt. 1: SPI slave mode CMD9 transmission is ended. 0: Others.

SLV_CMD9_INT_RAW_W

Field SLV_CMD9_INT_RAW writer - The raw bit for SPI slave CMD9 interrupt. 1: SPI slave mode CMD9 transmission is ended. 0: Others.

SLV_CMDA_INT_RAW_R

Field SLV_CMDA_INT_RAW reader - The raw bit for SPI slave CMDA interrupt. 1: SPI slave mode CMDA transmission is ended. 0: Others.

SLV_CMDA_INT_RAW_W

Field SLV_CMDA_INT_RAW writer - The raw bit for SPI slave CMDA interrupt. 1: SPI slave mode CMDA transmission is ended. 0: Others.

SLV_CMD_ERR_INT_RAW_R

Field SLV_CMD_ERR_INT_RAW reader - The raw bit for SPI_SLV_CMD_ERR_INT interrupt. 1: The slave command value in the current SPI slave HD mode transmission is not supported. 0: Others.

SLV_CMD_ERR_INT_RAW_W

Field SLV_CMD_ERR_INT_RAW writer - The raw bit for SPI_SLV_CMD_ERR_INT interrupt. 1: The slave command value in the current SPI slave HD mode transmission is not supported. 0: Others.

SLV_EN_QPI_INT_RAW_R

Field SLV_EN_QPI_INT_RAW reader - The raw bit for SPI slave En_QPI interrupt. 1: SPI slave mode En_QPI transmission is ended. 0: Others.

SLV_EN_QPI_INT_RAW_W

Field SLV_EN_QPI_INT_RAW writer - The raw bit for SPI slave En_QPI interrupt. 1: SPI slave mode En_QPI transmission is ended. 0: Others.

SLV_EX_QPI_INT_RAW_R

Field SLV_EX_QPI_INT_RAW reader - The raw bit for SPI slave Ex_QPI interrupt. 1: SPI slave mode Ex_QPI transmission is ended. 0: Others.

SLV_EX_QPI_INT_RAW_W

Field SLV_EX_QPI_INT_RAW writer - The raw bit for SPI slave Ex_QPI interrupt. 1: SPI slave mode Ex_QPI transmission is ended. 0: Others.

SLV_RD_BUF_DONE_INT_RAW_R

Field SLV_RD_BUF_DONE_INT_RAW reader - The raw bit for SPI_SLV_RD_BUF_DONE_INT interrupt. 1: SPI slave mode Rd_BUF transmission is ended. 0: Others.

SLV_RD_BUF_DONE_INT_RAW_W

Field SLV_RD_BUF_DONE_INT_RAW writer - The raw bit for SPI_SLV_RD_BUF_DONE_INT interrupt. 1: SPI slave mode Rd_BUF transmission is ended. 0: Others.

SLV_RD_DMA_DONE_INT_RAW_R

Field SLV_RD_DMA_DONE_INT_RAW reader - The raw bit for SPI_SLV_RD_DMA_DONE_INT interrupt. 1: SPI slave mode Rd_DMA transmission is ended. 0: Others.

SLV_RD_DMA_DONE_INT_RAW_W

Field SLV_RD_DMA_DONE_INT_RAW writer - The raw bit for SPI_SLV_RD_DMA_DONE_INT interrupt. 1: SPI slave mode Rd_DMA transmission is ended. 0: Others.

SLV_WR_BUF_DONE_INT_RAW_R

Field SLV_WR_BUF_DONE_INT_RAW reader - The raw bit for SPI_SLV_WR_BUF_DONE_INT interrupt. 1: SPI slave mode Wr_BUF transmission is ended. 0: Others.

SLV_WR_BUF_DONE_INT_RAW_W

Field SLV_WR_BUF_DONE_INT_RAW writer - The raw bit for SPI_SLV_WR_BUF_DONE_INT interrupt. 1: SPI slave mode Wr_BUF transmission is ended. 0: Others.

SLV_WR_DMA_DONE_INT_RAW_R

Field SLV_WR_DMA_DONE_INT_RAW reader - The raw bit for SPI_SLV_WR_DMA_DONE_INT interrupt. 1: SPI slave mode Wr_DMA transmission is ended. 0: Others.

SLV_WR_DMA_DONE_INT_RAW_W

Field SLV_WR_DMA_DONE_INT_RAW writer - The raw bit for SPI_SLV_WR_DMA_DONE_INT interrupt. 1: SPI slave mode Wr_DMA transmission is ended. 0: Others.

TRANS_DONE_INT_RAW_R

Field TRANS_DONE_INT_RAW reader - The raw bit for SPI_TRANS_DONE_INT interrupt. 1: SPI master mode transmission is ended. 0: others.

TRANS_DONE_INT_RAW_W

Field TRANS_DONE_INT_RAW writer - The raw bit for SPI_TRANS_DONE_INT interrupt. 1: SPI master mode transmission is ended. 0: others.

W

Register DMA_INT_RAW writer