Module esp32c2::system::sysclk_conf

system clock config register

Structs

SYSCLK_CONF_SPEC

system clock config register

Type Aliases

CLK_DIV_EN_R

Field CLK_DIV_EN reader - reg_clk_div_en

CLK_XTAL_FREQ_R

Field CLK_XTAL_FREQ reader - This field is used to read xtal frequency in MHz.

PRE_DIV_CNT_R

Field PRE_DIV_CNT reader - This field is used to set the count of prescaler of XTAL_CLK.

PRE_DIV_CNT_W

Field PRE_DIV_CNT writer - This field is used to set the count of prescaler of XTAL_CLK.

R

Register SYSCLK_CONF reader

SOC_CLK_SEL_R

Field SOC_CLK_SEL reader - This field is used to select soc clock.

SOC_CLK_SEL_W

Field SOC_CLK_SEL writer - This field is used to select soc clock.

W

Register SYSCLK_CONF writer